Added TimeManager implementation and changed a type in Simulation
Simulation::personCount uint16_t -> int

+++ src/Simulation/Simulation.h
@@ -20,7 +20,7 @@ namespace DeseaseSpreadSimulation
 		Desease desease;
 		bool withPrint = false;
 
-		uint16_t personCount = 1000;
+		unsigned int personCount = 1000;
 		std::vector<Person> persons;
 	};
 }
